Added a simple libary for parsing cookies
The scipt allows me to parse the Set-Cookie headers in the response of HTTP Headers.

The libary stores them and generates a list that can be parsed to the fetch without any complication (hopefully)

This libary should not be used for every webpage because its doesnt support any operation; it just supports nesecarry operations!